Mudando variável float para int

+++ 1-curso/secao 4: Estruturas logicas e condicionais /exercícios/ex3_par_ou_impar.py
@@ -1,7 +1,7 @@
 """3. Faça um programa que recebe um número inteiro e informe se este número é par ou ímpar."""
 
 
-NUMBER = float(input('Digite um número:'))
+NUMBER = int(input('Digite um número:'))
 
 if NUMBER % 2 == 0:
     print('O número é par')
